Various minor fixes
* Parallel finite differences
* fixed some types
* ADUPROP_MPI define to enable MPI

+++ test/covariance/main.cpp
@@ -24,7 +24,7 @@ int main(int argc, char* argv[]) {
 
   // Variable declaration
   int nbuses, nbranches, ngen, nload;
-  int tsteps = 20;
+  size_t tsteps = 20;
   size_t dim;
   pVector<double> xold, x, F;
   pVector<double> x0;
